NodeJs/challenges/desafio-03-findAfriend-api/src/use-cases/register-user.ts
import { User, UsersRepository } from '@/repositories/users-repository'
import { hash } from 'bcryptjs'
import { UserAlreadyExistsError } from './errors/user-already-exists'

interface RegisterUserUseCaseRequest {
  name: string
  email: string
  password: string
  role?: 'Admin' | 'Member'
}

interface RegisterUserUseCaseResponse {
  user: User
}

export class RegisterUserUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    name,
    email,
    password,
    role,
  }: RegisterUserUseCaseRequest): Promise<RegisterUserUseCaseResponse> {
    const password_hash = await hash(password, 6)

    const userWhitSameEmail = await this.usersRepository.findByEmail(email)

    if (userWhitSameEmail) {
      throw new UserAlreadyExistsError()
    }

    const user = await this.usersRepository.create({
      name,
      email,
      password_hash,
      role,
    })

    return { user }
  }
}
